Split virtual table into a dedicated header file
