+++ dev/storage.h
@@ -54,12 +54,13 @@
 #include "expression_object_type.h"
 #include "statement_serializer.h"
 #include "serializer_context.h"
-#include "schema/triggers.h"
 #include "object_from_column_builder.h"
 #include "row_extractor.h"
 #include "schema/table.h"
+#include "schema/virtual_table.h"
 #include "schema/column.h"
 #include "schema/index.h"
+#include "schema/triggers.h"
 #include "cte_storage.h"
 #include "util.h"
 #include "serializing_util.h"
