+++ include/sqlite_orm/sqlite_orm.h
@@ -12245,27 +12245,13 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 
 #ifndef SQLITE_ORM_IMPORT_STD_MODULE
 #include <string>  //  std::string
-#include <type_traits>  //  std::remove_const, std::is_member_pointer, std::true_type, std::false_type
+#include <type_traits>  //  std::remove_const, std::true_type, std::false_type
 #include <vector>  //  std::vector
-#include <tuple>  //  std::tuple_element
+#include <tuple>  //  std::tuple_element, std::make_tuple, std::get
 #include <utility>  //  std::forward, std::move
 #endif
 
 // #include "../functional/cxx_type_traits_polyfill.h"
-
-// #include "../functional/cxx_functional_polyfill.h"
-
-// #include "../functional/mpl.h"
-
-// #include "../functional/index_sequence_util.h"
-
-// #include "../tuple_helper/tuple_filter.h"
-
-// #include "../tuple_helper/tuple_traits.h"
-
-// #include "../tuple_helper/tuple_iteration.h"
-
-// #include "../tuple_helper/tuple_transformer.h"
 
 // #include "../member_traits/member_traits.h"
 
@@ -12300,139 +12286,6 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 // #include "../field_of.h"
 
 // #include "../constraints.h"
-
-// #include "column.h"
-
-namespace sqlite_orm::internal {
-
-    struct table_base {
-
-        /**
-         *  Table name.
-         */
-        std::string name;
-    };
-
-    /**
-     *  Mixin for fields of any mapped schema object, i.e. table or view, or any virtual or temporary table.
-     */
-    template<class... Cs>
-    struct mapped_columns_mixin {
-        using elements_type = std::tuple<Cs...>;
-
-        elements_type elements;
-
-        /*
-         *  Returns the number of elements of the specified type.
-         */
-        template<template<class...> class Trait>
-        static constexpr int count_of() {
-            using sequence_of = filter_tuple_sequence_t<elements_type, Trait>;
-            return int(sequence_of::size());
-        }
-
-        /*
-         *  Returns the number of columns having the specified constraint trait.
-         */
-        template<template<class...> class Trait>
-        static constexpr int count_of_columns_with() {
-            using col_index_sequence = col_index_sequence_with<elements_type, Trait>;
-            return int(col_index_sequence::size());
-        }
-
-        /*
-         *  Returns the number of columns having the specified constraint trait.
-         */
-        template<template<class...> class Trait>
-        static constexpr int count_of_columns_excluding() {
-            using excluded_col_index_sequence = col_index_sequence_excluding<elements_type, Trait>;
-            return int(excluded_col_index_sequence::size());
-        }
-
-        /**
-         *  Call passed lambda with all defined columns.
-         *  @param lambda Lambda called for each column. Function signature: `void(auto& column)`
-         */
-        template<class L>
-        void for_each_column(L&& lambda) const {
-            iterate_tuple(this->elements, col_index_sequence_of<elements_type>{}, lambda);
-        }
-
-        /**
-         *  Call passed lambda with columns not having the specified constraint trait `OpTrait`.
-         *  @param lambda Lambda called for each column.
-         */
-        template<template<class...> class OpTraitFn, class L>
-        void for_each_column_excluding(L&& lambda) const {
-            iterate_tuple(this->elements, col_index_sequence_excluding<elements_type, OpTraitFn>{}, lambda);
-        }
-
-        /**
-         *  Call passed lambda with columns not having the specified constraint trait `OpTrait`.
-         *  @param lambda Lambda called for each column.
-         */
-        template<class OpTraitQ, class L, satisfies<mpl::is_quoted_metafuntion, OpTraitQ> = true>
-        void for_each_column_excluding(L&& lambda) const {
-            this->template for_each_column_excluding<OpTraitQ::template fn>(lambda);
-        }
-
-        /**
-             *  Searches column name by class member pointer passed as the first argument.
-             *  @return column name or empty string if nothing found.
-             */
-        template<class M, satisfies<std::is_member_pointer, M> = true>
-        const std::string* find_column_name(M memberPointer) const {
-            using field_type = member_field_type_t<M>;
-
-            const std::string* res = nullptr;
-            iterate_tuple(this->elements,
-                          col_index_sequence_with_field_type<elements_type, field_type>{},
-                          [&res, memberPointer](auto& column) {
-                              if (compare_fields(column.member_pointer, memberPointer) ||
-                                  compare_fields(column.setter, memberPointer)) {
-                                  res = &column.name;
-                              }
-                          });
-            return res;
-        }
-    };
-
-    /**
-     *  Base for a mapped schema object aka table, view.
-     */
-    template<class O, class... Cs>
-    struct mapped_object_base : mapped_columns_mixin<Cs...> {
-        using base_type = mapped_columns_mixin<Cs...>;
-        using object_type = O;
-        using elements_type = typename base_type::elements_type;
-
-        /**
-         *  Function used to get field value from object by mapped member pointer/setter/getter.
-         *  
-         *  For a setter the corresponding getter has to be searched,
-         *  so the method returns a pointer to the field as returned by the found getter.
-         *  Otherwise the method invokes the member pointer and returns its result.
-         */
-        template<class M, satisfies_not<is_setter, M> = true>
-        decltype(auto) object_field_value(const object_type& object, M memberPointer) const {
-            return polyfill::invoke(memberPointer, object);
-        }
-
-        template<class M, satisfies<is_setter, M> = true>
-        const member_field_type_t<M>* object_field_value(const object_type& object, M memberPointer) const {
-            using field_type = member_field_type_t<M>;
-            const field_type* res = nullptr;
-            iterate_tuple(this->elements,
-                          col_index_sequence_with_field_type<elements_type, field_type>{},
-                          call_as_template_base<column_field>([&res, &memberPointer, &object](const auto& column) {
-                              if (compare_fields(column.setter, memberPointer)) {
-                                  res = &polyfill::invoke(column.member_pointer, object);
-                              }
-                          }));
-            return res;
-        }
-    };
-}
 
 // #include "index.h"
 
@@ -12570,22 +12423,155 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 
 // #include "column.h"
 
+namespace sqlite_orm::internal {
+
+    template<class T>
+    using is_table_element_or_constraint = mpl::invoke_t<mpl::disjunction<check_if<is_column>,
+                                                                          check_if<is_primary_key>,
+                                                                          check_if<is_foreign_key>,
+                                                                          check_if_is_template<index_t>,
+                                                                          check_if_is_template<unique_t>,
+                                                                          check_if_is_template<check_t>,
+                                                                          check_if_is_template<prefix_t>,
+                                                                          check_if_is_template<tokenize_t>,
+                                                                          check_if_is_template<content_t>,
+                                                                          check_if_is_template<table_content_t>>,
+                                                         T>;
+
+    struct table_base {
+
+        /**
+         *  Table name.
+         */
+        std::string name;
+    };
+
+    /**
+     *  Mixin for fields of any mapped schema object, i.e. table or view, or any virtual or temporary table.
+     */
+    template<class... Cs>
+    struct mapped_columns_mixin {
+        using elements_type = std::tuple<Cs...>;
+
+        elements_type elements;
+
+        /*
+         *  Returns the number of elements of the specified type.
+         */
+        template<template<class...> class Trait>
+        static constexpr int count_of() {
+            using sequence_of = filter_tuple_sequence_t<elements_type, Trait>;
+            return int(sequence_of::size());
+        }
+
+        /*
+         *  Returns the number of columns having the specified constraint trait.
+         */
+        template<template<class...> class Trait>
+        static constexpr int count_of_columns_with() {
+            using col_index_sequence = col_index_sequence_with<elements_type, Trait>;
+            return int(col_index_sequence::size());
+        }
+
+        /*
+         *  Returns the number of columns having the specified constraint trait.
+         */
+        template<template<class...> class Trait>
+        static constexpr int count_of_columns_excluding() {
+            using excluded_col_index_sequence = col_index_sequence_excluding<elements_type, Trait>;
+            return int(excluded_col_index_sequence::size());
+        }
+
+        /**
+         *  Call passed lambda with all defined columns.
+         *  @param lambda Lambda called for each column. Function signature: `void(auto& column)`
+         */
+        template<class L>
+        void for_each_column(L&& lambda) const {
+            iterate_tuple(this->elements, col_index_sequence_of<elements_type>{}, lambda);
+        }
+
+        /**
+         *  Call passed lambda with columns not having the specified constraint trait `OpTrait`.
+         *  @param lambda Lambda called for each column.
+         */
+        template<template<class...> class OpTraitFn, class L>
+        void for_each_column_excluding(L&& lambda) const {
+            iterate_tuple(this->elements, col_index_sequence_excluding<elements_type, OpTraitFn>{}, lambda);
+        }
+
+        /**
+         *  Call passed lambda with columns not having the specified constraint trait `OpTrait`.
+         *  @param lambda Lambda called for each column.
+         */
+        template<class OpTraitQ, class L, satisfies<mpl::is_quoted_metafuntion, OpTraitQ> = true>
+        void for_each_column_excluding(L&& lambda) const {
+            this->template for_each_column_excluding<OpTraitQ::template fn>(lambda);
+        }
+
+        /**
+             *  Searches column name by class member pointer passed as the first argument.
+             *  @return column name or empty string if nothing found.
+             */
+        template<class M, satisfies<std::is_member_pointer, M> = true>
+        const std::string* find_column_name(M memberPointer) const {
+            using field_type = member_field_type_t<M>;
+
+            const std::string* res = nullptr;
+            iterate_tuple(this->elements,
+                          col_index_sequence_with_field_type<elements_type, field_type>{},
+                          [&res, memberPointer](auto& column) {
+                              if (compare_fields(column.member_pointer, memberPointer) ||
+                                  compare_fields(column.setter, memberPointer)) {
+                                  res = &column.name;
+                              }
+                          });
+            return res;
+        }
+    };
+
+    /**
+     *  Base for a mapped schema object aka table, view.
+     */
+    template<class O, class... Cs>
+    struct mapped_object_base : mapped_columns_mixin<Cs...> {
+        using base_type = mapped_columns_mixin<Cs...>;
+        using object_type = O;
+        using elements_type = typename base_type::elements_type;
+
+        /**
+         *  Function used to get field value from object by mapped member pointer/setter/getter.
+         *  
+         *  For a setter the corresponding getter has to be searched,
+         *  so the method returns a pointer to the field as returned by the found getter.
+         *  Otherwise the method invokes the member pointer and returns its result.
+         */
+        template<class M, satisfies_not<is_setter, M> = true>
+        decltype(auto) object_field_value(const object_type& object, M memberPointer) const {
+            return polyfill::invoke(memberPointer, object);
+        }
+
+        template<class M, satisfies<is_setter, M> = true>
+        const member_field_type_t<M>* object_field_value(const object_type& object, M memberPointer) const {
+            using field_type = member_field_type_t<M>;
+            const field_type* res = nullptr;
+            iterate_tuple(this->elements,
+                          col_index_sequence_with_field_type<elements_type, field_type>{},
+                          call_as_template_base<column_field>([&res, &memberPointer, &object](const auto& column) {
+                              if (compare_fields(column.setter, memberPointer)) {
+                                  res = &polyfill::invoke(column.member_pointer, object);
+                              }
+                          }));
+            return res;
+        }
+    };
+}
+
+// #include "column.h"
+
 namespace sqlite_orm {
 
     namespace internal {
-
-        template<class T>
-        using is_table_element_or_constraint = mpl::invoke_t<mpl::disjunction<check_if<is_column>,
-                                                                              check_if<is_primary_key>,
-                                                                              check_if<is_foreign_key>,
-                                                                              check_if_is_template<index_t>,
-                                                                              check_if_is_template<unique_t>,
-                                                                              check_if_is_template<check_t>,
-                                                                              check_if_is_template<prefix_t>,
-                                                                              check_if_is_template<tokenize_t>,
-                                                                              check_if_is_template<content_t>,
-                                                                              check_if_is_template<table_content_t>>,
-                                                             T>;
 
         /**
          *  Table definition.
@@ -12706,29 +12692,6 @@ namespace sqlite_orm {
         template<class O, bool W, class... Cs>
         struct is_table<table_t<O, W, Cs...>> : std::true_type {};
 
-        template<class M>
-        struct virtual_table : table_base, M {
-            using module_type = M;
-            using object_type = typename module_type::object_type;
-            using elements_type = typename module_type::elements_type;
-
-            static constexpr bool is_without_rowid_v = false;
-            using is_without_rowid = polyfill::bool_constant<is_without_rowid_v>;
-
-            const module_type& module() const {
-                return *this;
-            }
-        };
-
-#if SQLITE_VERSION_NUMBER >= 3009000
-        template<class T, class... Cs>
-        struct fts5_module : mapped_columns_mixin<Cs...> {
-            using base_type = mapped_columns_mixin<Cs...>;
-            using object_type = T;
-            using elements_type = typename base_type::elements_type;
-        };
-#endif
-
         template<class O, bool WithoutRowId, class... Cs, class G, class S>
         bool exists_in_composite_primary_key(const table_t<O, WithoutRowId, Cs...>& table,
                                              const column_field<G, S>& column) {
@@ -12748,34 +12711,10 @@ namespace sqlite_orm {
             });
             return res;
         }
-
-        template<class M, class G, class S>
-        bool exists_in_composite_primary_key(const virtual_table<M>& /*virtualTable*/,
-                                             const column_field<G, S>& /*column*/) {
-            return false;
-        }
     }
 }
 
 SQLITE_ORM_EXPORT namespace sqlite_orm {
-#if SQLITE_VERSION_NUMBER >= 3009000
-    template<class... Cs, class T = typename std::tuple_element_t<0, std::tuple<Cs...>>::object_type>
-    internal::fts5_module<T, Cs...> using_fts5(Cs... columns) {
-        static_assert(polyfill::conjunction_v<internal::is_table_element_or_constraint<Cs>...>,
-                      "Incorrect table elements or constraints");
-
-        SQLITE_ORM_CLANG_SUPPRESS_MISSING_BRACES(return {std::make_tuple(std::forward<Cs>(columns)...)});
-    }
-
-    template<class T, class... Cs>
-    internal::fts5_module<T, Cs...> using_fts5(Cs... columns) {
-        static_assert(polyfill::conjunction_v<internal::is_table_element_or_constraint<Cs>...>,
-                      "Incorrect table elements or constraints");
-
-        SQLITE_ORM_CLANG_SUPPRESS_MISSING_BRACES(return {std::make_tuple(std::forward<Cs>(columns)...)});
-    }
-#endif
-
     /**
      *  Factory function for a table definition.
      *  
@@ -12815,11 +12754,6 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
         return make_table<internal::auto_decay_table_ref_t<table>>(std::move(name), std::forward<Cs>(args)...);
     }
 #endif
-
-    template<class M>
-    internal::virtual_table<M> make_virtual_table(std::string name, M module) {
-        SQLITE_ORM_CLANG_SUPPRESS_MISSING_BRACES(return {std::move(name), std::move(module)});
-    }
 }
 
 // #include "storage_lookup.h"
@@ -20266,6 +20200,77 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 
 // #include "schema/table.h"
 
+// #include "schema/virtual_table.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <string>  //  std::string
+#include <tuple>  //  std::tuple_element, std::make_tuple
+#include <utility>  //  std::forward, std::move
+#endif
+
+// #include "../functional/cxx_type_traits_polyfill.h"
+
+// #include "table_base.h"
+
+// #include "column.h"
+
+namespace sqlite_orm::internal {
+
+    template<class M>
+    struct virtual_table : table_base, M {
+        using module_type = M;
+        using object_type = typename module_type::object_type;
+        using elements_type = typename module_type::elements_type;
+
+        static constexpr bool is_without_rowid_v = false;
+        using is_without_rowid = polyfill::bool_constant<is_without_rowid_v>;
+
+        const module_type& module() const {
+            return *this;
+        }
+    };
+
+#if SQLITE_VERSION_NUMBER >= 3009000
+    template<class T, class... Cs>
+    struct fts5_module : mapped_columns_mixin<Cs...> {
+        using base_type = mapped_columns_mixin<Cs...>;
+        using object_type = T;
+        using elements_type = typename base_type::elements_type;
+    };
+#endif
+
+    template<class M, class G, class S>
+    bool exists_in_composite_primary_key(const virtual_table<M>& /*virtualTable*/,
+                                         const column_field<G, S>& /*column*/) {
+        return false;
+    }
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+#if SQLITE_VERSION_NUMBER >= 3009000
+    template<class... Cs, class T = typename std::tuple_element_t<0, std::tuple<Cs...>>::object_type>
+    internal::fts5_module<T, Cs...> using_fts5(Cs... columns) {
+        static_assert(polyfill::conjunction_v<internal::is_table_element_or_constraint<Cs>...>,
+                      "Incorrect table elements or constraints");
+
+        SQLITE_ORM_CLANG_SUPPRESS_MISSING_BRACES(return {std::make_tuple(std::forward<Cs>(columns)...)});
+    }
+
+    template<class T, class... Cs>
+    internal::fts5_module<T, Cs...> using_fts5(Cs... columns) {
+        static_assert(polyfill::conjunction_v<internal::is_table_element_or_constraint<Cs>...>,
+                      "Incorrect table elements or constraints");
+
+        SQLITE_ORM_CLANG_SUPPRESS_MISSING_BRACES(return {std::make_tuple(std::forward<Cs>(columns)...)});
+    }
+#endif
+
+    template<class M>
+    internal::virtual_table<M> make_virtual_table(std::string name, M module) {
+        SQLITE_ORM_CLANG_SUPPRESS_MISSING_BRACES(return {std::move(name), std::move(module)});
+    }
+}
+
 namespace sqlite_orm {
 
     namespace internal {
@@ -22680,17 +22685,19 @@ namespace sqlite_orm {
 
 // #include "serializer_context.h"
 
-// #include "schema/triggers.h"
-
 // #include "object_from_column_builder.h"
 
 // #include "row_extractor.h"
 
 // #include "schema/table.h"
 
+// #include "schema/virtual_table.h"
+
 // #include "schema/column.h"
 
 // #include "schema/index.h"
+
+// #include "schema/triggers.h"
 
 // #include "cte_storage.h"
